pyproject.toml
[tool.poetry]
name = "notion-nlp"
version = "1.0.6"
description = "Reading rich text information from a Notion database and performing simple NLP analysis."
authors = ["Dario Zhang <zdclink@gmail.com>"]
license = "MIT license"
readme = "README.md"
packages = [{include = "notion_nlp", from = "src"}]

[tool.poetry.dependencies]
python = "^3.8.10,<3.12"
arrow = "1.2.3"
pyfunctional = "1.4.3"
ruamel-yaml = "0.17.21"
tqdm = "4.64.1"
pandas = "1.5.3"
scikit-learn = "1.2.1"
jieba = "0.42.1"
tabulate = "0.9.0"
wcwidth = "0.2.6"
requests = "2.28.2"
typer = "0.7.0"
pydantic = "1.10.5"
wordcloud = "1.8.2.2"
matplotlib = "3.7.0"

[tool.poetry.group.dev.dependencies]
py = "1.11.0"
pytest = "7.2.1"
pytest-cover = "3.0.0"
pytest-html = "3.1.1"
black = "23.1.0"
isort = "5.12.0"
flake8 = "6.0.0"
pylint = "2.16.2"
tox = "3.28.0"
invoke = "1.7.3"
setuptools = "67.3.2"
sphinx = {version = "3.4.0", optional = true}
sphinx-rtd-theme = {version = "0.5.0", optional = true}
recommonmark = {version = "0.7.1", optional = true}
sphinx-markdown-tables = {version = "0.0.15", optional = true}
pyinstaller = "^5.8.0"
memory-profiler = "^0.61.0"


[tool.poetry.extras]
doc = ["sphinx", "sphinx-rtd-theme", "recommonmark", "sphinx-markdown-tables"]

[tool.black]
line-length = 90

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 90
known_third_party = ""
skip = "test,tests"

[tool.coverage.run]
# include source files
source=["./src/"]
# exclude source files
omit=[]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
